+++ src/extension.js
@@ -24,6 +24,10 @@ function inspect(input, filepath, parseImports){
     typeof parseImports == "undefined" || parseImports==false ? false : true
 
     console.log("inspect: " + filepath)
+    if (!fs.existsSync(filepath)){
+        console.error("file does not exist! --> " + filepath)
+        return
+    }
     /** cachelookup first */
     let hash = crypto.createHash('sha1').update(input).digest('base64')
     
@@ -46,9 +50,30 @@ function inspect(input, filepath, parseImports){
 
     /** parse imports */
     sourceUnit.imports.forEach(function(imp){
-
+        //try file path
         let importPath = path.resolve(path.dirname(filepath) +"/./" + imp.path)
-        imp.ast = inspectFile(importPath, true)  // this caches automatically
+        if (!fs.existsSync(importPath)){
+            //try relative to workspace root
+            importPath = path.resolve(vscode.workspace.rootPath +"/./" + imp.path)
+            if (!fs.existsSync(importPath)){
+                //try workspacepath node_modules
+                importPath = path.resolve(vscode.workspace.rootPath +"/node_modules/" + imp.path)
+                if (!fs.existsSync(importPath)){
+                    //try ../contracts/node_modules/...
+                    let basepath = filepath.split("/contracts/")
+
+                    if (basepath.length==2){ //super dirty
+                        basepath=basepath[0]
+                        importPath = path.resolve(basepath +"/node_modules/" + imp.path)
+                    }
+                }
+            }
+        }
+        try {
+            imp.ast = inspectFile(importPath, true)  // this caches automatically
+        } catch (e) {
+            console.error(e)
+        }
     })
     /** we're done */
     // should we flatten inherited functions to one object?
@@ -105,15 +130,16 @@ function parseSourceUnit(input){
                 _node: node,
                 name: node.name,
                 dependencies: node.baseContracts.map(spec => spec.baseName.namePath),
-                stateVars:{},
-                enums:{},
-                structs:{},
-                mappings:{},
-                modifiers:{},
-                functions:{},
-                constructor:null,
-                events:{},
-                inherited_names: {}
+                stateVars:{},  // pure statevars --> see names
+                enums:{},  // enum declarations
+                structs:{}, // struct declarations
+                mappings:{},  // mapping declarations
+                modifiers:{},  // modifier declarations 
+                functions:{},  // function and method declarations
+                constructor:null,  // ...
+                events:{},  // event declarations
+                inherited_names: {},  // all names inherited from other contracts
+                names:{}   // all names in current contract (methods, events, structs, ...)
             }
             current_contract = sourceUnit.contracts[node.name]
 
@@ -124,13 +150,23 @@ function parseSourceUnit(input){
                         VariableDeclaration(__node){
                             __node.usedAt = new Array();
                             current_contract.stateVars[__node.name]=__node;
+                            current_contract.names[__node.name]=__node;
                         }
                     })
                 },
                 // --> is a subtype. Mapping(_node){current_contract.mappings[_node.name]=_node},
-                EnumDefinition(_node){current_contract.enums[_node.name]=_node},
-                StructDefinition(_node){current_contract.structs[_node.name]=_node},
-                ConstructorDefinition(_node){current_contract.constructor=_node}, // wrong def in code: https://github.com/solidityj/solidity-antlr4/blob/fbe865f8ba510cbdb1540fcf9517a42820a4d097/Solidity.g4#L78 for consttuctzor () ..
+                EnumDefinition(_node){
+                    current_contract.enums[_node.name]=_node
+                    current_contract.names[_node.name]=_node
+                },
+                StructDefinition(_node){
+                    current_contract.structs[_node.name]=_node
+                    current_contract.names[_node.name]=_node
+                },
+                ConstructorDefinition(_node){
+                    current_contract.constructor=_node
+                    current_contract.names[_node.name]=_node
+                }, // wrong def in code: https://github.com/solidityj/solidity-antlr4/blob/fbe865f8ba510cbdb1540fcf9517a42820a4d097/Solidity.g4#L78 for consttuctzor () ..
                 ModifierDefinition(_node){
                     current_contract.modifiers[_node.name]={
                         _node:_node,
@@ -139,7 +175,9 @@ function parseSourceUnit(input){
                         declarations: {},  // all declarations: arguments+returns+body
                         identifiers: [],  // all identifiers (use of variables)
                     }
+                    
                     current_function = current_contract.modifiers[_node.name];
+                    current_contract.names[_node.name]=current_function;
                     // parse function body to get all function scope params.
                     // first get declarations
                     parser.visit(_node.parameters, {
@@ -173,8 +211,27 @@ function parseSourceUnit(input){
                         }
                     })
                 },
-                EventDefinition(_node){current_contract.events[_node.name]=_node},
+                EventDefinition(_node){
+                    current_contract.events[_node.name]={
+                        _node:_node,
+                        arguments: {},  // declarations: quick access to argument list
+                        declarations: {},  // all declarations: arguments+returns+body
+                    }
+                    
+                    current_function = current_contract.events[_node.name];
+                    current_contract.names[_node.name]=current_function;
+                    // parse function body to get all function scope params.
+                    // first get declarations
+                    parser.visit(_node.parameters, {
+                        VariableDeclaration: function(__node){
+                            current_function.arguments[__node.name]=__node
+                            current_function.declarations[__node.name]=__node
+                        }
+                    })
+
+                },
                 FunctionDefinition(_node){
+                    
                     current_contract.functions[_node.name]={
                         _node:_node,
                         arguments: {},  // declarations: quick access to argument list
@@ -185,6 +242,7 @@ function parseSourceUnit(input){
                         accesses_svar: false //
                     }
                     current_function = current_contract.functions[_node.name];
+                    current_contract.names[_node.name]=current_function;
                     // parse function body to get all function scope params.
                     // first get declarations
                     parser.visit(_node.parameters, {
@@ -258,7 +316,7 @@ function parseSourceUnit(input){
                 
                 if(typeof sourceUnit.contracts[contract].stateVars[identifier.name]!="undefined"){
                     sourceUnit.contracts[contract].stateVars[identifier.name].usedAt.push(identifier)
-                    func.accesses_svar = true;
+                    func.accesses_svar = true;  // TODO: also check for inherited svars
                 }
                 
                 for (var identDec in func.arguments){
@@ -308,7 +366,10 @@ function linearizeContract(sourceUnit){
         dependencies[contractName] = sourceUnit.contracts[contractName].dependencies;
         sourceUnit.contracts[contractName].dependencies.forEach(function(dep){
             let contract = g_contracts[dep]
-            dependencies[dep]=contract.dependencies;
+            if(typeof contract=="undefined")
+                console.error("ERROR - could not load contract object for "+dep)
+            else
+                dependencies[dep]=contract.dependencies;
         })
     }
     return linearize(dependencies, {reverse: true})
@@ -324,12 +385,12 @@ function onDidChange(event){
     console.log("did-change")
     //mod_decorator.updateDecorations();
     console.log("inspect")
+ 
     var insights = inspect(activeEditor.document.getText(), activeEditor.document.fileName);
     console.log("inspect - end")
 
     console.log("linearize")
     var inheritance = linearizeContract(insights)
-    console.log(inheritance)
     console.log("linearize - end")
 
     var words = new Array();
@@ -342,6 +403,10 @@ function onDidChange(event){
         //merge all contracts into one
         inheritance[contract].forEach(function(contractName){
             var subcontract = g_contracts[contractName];
+            if(typeof subcontract=="undefined"){
+                console.error("ERROR - contract object not available "+ contractName)
+                return
+            }
 
             for (let _var in subcontract.stateVars){
                 if(subcontract.stateVars[_var].visibility!="private")
@@ -372,8 +437,9 @@ function onDidChange(event){
                     insights.contracts[contract].inherited_names[_var] = contractName;
             }
         })
-        console.log(insights.contracts[contract].inherited_names)
         console.log("resolve inheritance -- done")
+
+        /** todo fixme: rework */
         for (var stateVar in insights.contracts[contract].stateVars) {
             let svar = insights.contracts[contract].stateVars[stateVar];
             // only statevars that are not const
@@ -403,7 +469,7 @@ function onDidChange(event){
             });
             
 
-            console.log("--annoate idents--")
+            //console.log("--annoate idents--")
             /*** annotate all identifiers */
             //console.log(svar.usedAt)
             svar.usedAt.forEach(function (ident){
@@ -445,11 +511,95 @@ function onDidChange(event){
                 }
             })
         }
+
+        /*** inherited vars */
+        /** have to get all identifiers :// */
+        /** fixme ugly hack ... */
+        for (var functionName in insights.contracts[contract].functions){
+            //all functions
+            insights.contracts[contract].functions[functionName].identifiers.forEach(function(ident){
+                // all idents in function
+                let is_state_var = typeof insights.contracts[contract].stateVars[ident.name]!="undefined"
+                let is_declared_locally = typeof ident.inFunction.declarations[ident.name]!="undefined"
+                let is_inherited = typeof insights.contracts[contract].inherited_names[ident.name]!="undefined" && insights.contracts[contract].inherited_names[ident.name]!=contract
+
+                if(is_declared_locally){
+                    if(is_state_var){
+                        //shadowed staevar
+                        console.log("!!!! shadowed statevar")
+                        //is handled in the other loop
+                    }else if(is_inherited){
+                        //shadoewed inherited var
+                        console.log("!!!!! shadowed derived var")
+                        prefix = "**INHERITED**  ❗SHADOWED❗"
+                        decoStyle = "decoStyleLightOrange";
+                        let subcontract =  insights.contracts[contract].inherited_names[ident.name]
+                        var decl_uri = "([Declaration: #"+(ident.loc.start.line)+"]("+activeEditor.document.uri+":"+(ident.loc.start.line)+":1))"
+
+                        decorations.push(
+                            { 
+                                range: new vscode.Range(
+                                    new vscode.Position(ident.loc.start.line-1, ident.loc.start.column),
+                                    new vscode.Position(ident.loc.end.line-1, ident.loc.end.column+ident.name.length)
+                                    ),
+                                    hoverMessage: prefix + "(*"+ "undef" +"*) " +'**StateVar** *' + subcontract + "*.**"+ident.name + '**' + " "+decl_uri,
+                                decoStyle: decoStyle
+                            });
+                    }else {
+                        //all good
+                        // is declared locally
+                    }
+                } else if(is_state_var){
+                    if(is_inherited){
+                        //shadowed inherited var
+                        console.log("!!! statevar shadows inherited")
+                        console.log("!!!!! shadowed derived var")
+                        prefix = "**INHERITED**  ❗SHADOWED❗"
+                        decoStyle = "decoStyleLightOrange";
+                        let subcontract =  insights.contracts[contract].inherited_names[ident.name]
+                        var decl_uri = "([Declaration: #"+(ident.loc.start.line)+"]("+activeEditor.document.uri+":"+(ident.loc.start.line)+":1))"
+
+                        decorations.push(
+                            { 
+                                range: new vscode.Range(
+                                    new vscode.Position(ident.loc.start.line-1, ident.loc.start.column),
+                                    new vscode.Position(ident.loc.end.line-1, ident.loc.end.column+ident.name.length)
+                                    ),
+                                    hoverMessage: prefix + "(*"+ "undef" +"*) " +'**StateVar** *' + subcontract + "*.**"+ident.name + '**' + " "+decl_uri,
+                                decoStyle: decoStyle
+                            });
+                    } else {
+                        //all good statevar
+                        // should be covered by the other loop already
+                    }
+                } else if (is_inherited){
+                    // inherited
+                    prefix = "**INHERITED**  "
+                    decoStyle = "decoStyleLightBlue";
+                    let subcontract =  insights.contracts[contract].inherited_names[ident.name]
+                    var decl_uri = "([Declaration: #"+(ident.loc.start.line)+"]("+activeEditor.document.uri+":"+(ident.loc.start.line)+":1))"
+
+                    decorations.push(
+                        { 
+                            range: new vscode.Range(
+                                new vscode.Position(ident.loc.start.line-1, ident.loc.start.column),
+                                new vscode.Position(ident.loc.end.line-1, ident.loc.end.column+ident.name.length)
+                                ),
+                                hoverMessage: prefix + "(*"+ "undef" +"*) " +'**StateVar** *' + subcontract + "*.**"+ident.name + '**' + " "+decl_uri,
+                            decoStyle: decoStyle
+                        });
+                } else {
+                    //function calls etc.. fallthru
+                    
+                }
+            })
+        }
     }
     console.log("--set-decorate--")
     //mod_decorator.decorateWords(activeEditor, words);
     if (solidityVAConfig.deco.statevars)
         setDecorations(activeEditor, decorations)
+
     console.log("deco end")
 }
 
@@ -472,7 +622,7 @@ function setDecorations(editor, decorations){
     }
 }
 
-function getSymbolForNode(document, node, kind, name, containerName){
+function getSymbolForNode(document, node, kind, name, detail){
     let range = new vscode.Range(
         new vscode.Position(node.loc.start.line-1, node.loc.start.column),
         new vscode.Position(node.loc.end.line-1, typeof node.length!="undefined"?node.loc.end.column+node.length:node.loc.end.column)
@@ -480,7 +630,9 @@ function getSymbolForNode(document, node, kind, name, containerName){
     return {
         name: typeof name!="undefined"?name:node.name,
         kind: kind,
-        location: new vscode.Location(document.uri, range),
+        detail: typeof detail!="undefined"?detail:"",
+        location: new vscode.Location(document.uri, range)
+        //children: []
     }
 }
 
@@ -527,122 +679,418 @@ function getVisibilityToIcon(state){
     return v
 }
 
+function astNodeAsDocumentSymbol(document, node, kind, name, detail){
+    return new vscode.DocumentSymbol(
+        typeof name!="undefined"?name:node.name, 
+        typeof detail!="undefined"?detail:"", 
+        kind,
+        new vscode.Range(
+            node.loc.start.line-1,
+            node.loc.start.column,
+            node.loc.end.line-1, 
+            typeof node.length!="undefined"?node.loc.end.column+node.length:node.loc.end.column
+            ), 
+        new vscode.Range(
+            node.loc.start.line-1,
+            node.loc.start.column,
+            node.loc.end.line-1, 
+            typeof node.length!="undefined"?node.loc.end.column+node.length:node.loc.end.column
+            )
+    )
+    
+}
+
+function varDecIsArray(node){
+    return node.typeName.type=="ArrayTypeName"
+}
+
+function varDecIsUserDefined(node){
+    return node.typeName.type=="UserDefinedTypeName"
+}
+
+function getVariableDeclarationType(node){
+    if(varDecIsArray(node))
+        node = node.typeName.baseTypeName
+    else
+        node = node.typeName
+    if(node.type=="ElementaryTypeName"){
+        return node.name;
+    } else if (node.type=="UserDefinedTypeName"){
+        return node.namePath;
+    } else {
+        return null
+    }
+}
+
+const varDecToSymbolType = {
+    string: vscode.SymbolKind.String,
+    array: vscode.SymbolKind.Array,
+    bool: vscode.SymbolKind.Boolean,
+    uint: vscode.SymbolKind.Number,
+    int: vscode.SymbolKind.Number,
+    bytes: vscode.SymbolKind.Array,
+    address: vscode.SymbolKind.Variable,
+    user: vscode.SymbolKind.Array
+}
+
+function getVariableDeclarationTypeAsSymbolKind(node, _default){
+    if(varDecIsUserDefined(node)){
+        return varDecToSymbolType.user
+    }
+
+    let solidityType = getVariableDeclarationType(node)
+    if (!solidityType)
+        return _default
+        
+    if (solidityType.startsWith("uint") || solidityType.startsWith("int")){
+        return varDecToSymbolType.uint;
+    } else if (solidityType.startsWith("bytes")){
+        return varDecToSymbolType.bytes;
+    }
+
+    let kind = varDecToSymbolType[solidityType]
+    return typeof kind!="undefined"?kind:_default;
+    
+}
+
+function getSymbolKindForDeclaration(node){
+    let astnode = typeof node._node!="undefined"?node._node:node
+    let result = {
+        symbol: vscode.SymbolKind.Variable,
+        prefix: "",
+        suffix: "",
+        name: typeof astnode.name!="undefined" ? astnode.name:"",
+        details: ""
+    }
+    switch(astnode.type) {
+        case "ModifierDefinition":
+            result.symbol = vscode.SymbolKind.Method 
+
+            result.prefix += "Ⓜ "
+
+            if (solidityVAConfig.outline.decorations){
+                result.prefix += getVisibilityToIcon(astnode.visibility)
+                result.prefix += getStateMutabilityToIcon(astnode.stateMutability)
+            }
+            break;
+        case "EventDefinition":
+            result.symbol = vscode.SymbolKind.Event
+            break;
+        case "FunctionDefinition":
+            if (astnode.isConstructor){
+                result.symbol = vscode.SymbolKind.Constructor
+                result.name = "⚜ __constructor__ " + result.name
+            } else {
+                result.symbol = vscode.SymbolKind.Function
+            }
+            if (solidityVAConfig.outline.decorations){
+                result.prefix += getVisibilityToIcon(astnode.visibility)
+                result.prefix += getStateMutabilityToIcon(astnode.stateMutability)
+            }
+
+            if (solidityVAConfig.outline.extras){
+                result.suffix += " ( "
+                result.suffix += " complex: "+node.complexity
+                result.suffix += " state: " + (node.accesses_svar?"☑":"☐")
+                result.suffix += " )"
+            }
+            break;
+        case "EnumDefinition":
+            result.symbol = vscode.SymbolKind.Enum
+            break;
+        case "StructDefinition":
+            result.symbol = vscode.SymbolKind.Struct
+            break;
+        case "VariableDeclaration":
+            if(solidityVAConfig.outline.var.storage_annotations){
+                if(astnode.storageLocation=="memory"){
+                    result.prefix +="💾"
+                    result.details += astnode.storageLocation
+                } else if(astnode.storageLocation=="storage"){
+                    result.prefix +="💽"
+                    result.details += astnode.storageLocation
+                }
+            }
+            if(varDecIsArray(astnode)){
+                result.name += "[]"
+            }
+
+            if(astnode.isDeclaredConst){
+                result.symbol = vscode.SymbolKind.Constant
+                break
+            }
+            //result.symbol = vscode.SymbolKind.Variable
+
+            result.symbol = getVariableDeclarationTypeAsSymbolKind(astnode, vscode.SymbolKind.Variable)
+            break;
+        case "Parameter":
+            if(solidityVAConfig.outline.var.storage_annotations){
+                if(astnode.storageLocation=="memory"){
+                    result.prefix +="💾"
+                    result.details += astnode.storageLocation
+                } else if(astnode.storageLocation=="storage"){
+                    result.prefix +="💽"
+                    result.details += astnode.storageLocation
+                }
+            }
+            if(varDecIsArray(astnode)){
+                result.name += "[]"
+            }
+            result.symbol = vscode.SymbolKind.TypeParameter  // lets misuse this kind for params
+            break;
+        default:
+            console.log(node)
+            console.log("<-----")
+    }
+    return result;
+}  
+
 class SolidityDocumentSymbolProvider{
     provideDocumentSymbols(document, token){
+        console.log("force onDidChange() event to make sure ast is ready when providing symbols")
+        onDidChange()
+        onDidChange()
         console.log("preparing symbols...")
         return new Promise((resolve, reject) => {
             var symbols = [];
             var insights = inspect(document.getText());
+            console.log("preparing symbols for: "+ document.fileName)
 
-            /*
-            symbols.push(
-                getSymbolForNode(
+            if(solidityVAConfig.outline.pragmas.show){
+                var topLevelNode = astNodeAsDocumentSymbol(
                     document, 
-                    getFakeNode(path.basename(document.fileName),1),
-                    vscode.SymbolKind.File
+                    getFakeNode("pragma",1),
+                    vscode.SymbolKind.Namespace,
+                    "pragma",
+                    "... (" + insights.pragmas.length + ")"
                     )
-                )
-            */
-            
-            /*
-            insights.imports.forEach(function(imp){
-                imp.symbolAliases.forEach(function(sa){
-                    sa.forEach(function(impname){
-                        if(impname!=null)
-                            symbols.push(getSymbolForNode(document, imp, vscode.SymbolKind.Namespace, impname, "_imports_"))
+                symbols.push(topLevelNode)
+                insights.pragmas.forEach(function(item){
+                    topLevelNode.children.push(astNodeAsDocumentSymbol(
+                        document, 
+                        item, 
+                        vscode.SymbolKind.Namespace, 
+                        item.name + " → " + item.value))
                     })
-                })
-            })
-            */
+                    console.log("✓ pragmas ")
+            }
             
-
+            if(solidityVAConfig.outline.imports.show){
+                topLevelNode = astNodeAsDocumentSymbol(
+                    document, 
+                    getFakeNode("imports",1),
+                    vscode.SymbolKind.Namespace,
+                    "imports",
+                    "... (" + insights.imports.length + ")"
+                    )
+                symbols.push(topLevelNode)
+                insights.imports.forEach(function(item){
+                    topLevelNode.children.push(astNodeAsDocumentSymbol(
+                        document, 
+                        item, 
+                        vscode.SymbolKind.File, 
+                        item.path))
+                    })
+                
+                console.log("✓ imports ")
+            }
+            
             for (var contractName in insights.contracts) {
                 
-                
+                topLevelNode = astNodeAsDocumentSymbol(document, insights.contracts[contractName]._node, vscode.SymbolKind.Class)
+                symbols.push(topLevelNode)
+
                 /** the document */
-                symbols.push(getSymbolForNode(document, insights.contracts[contractName]._node, vscode.SymbolKind.Class))
+                console.log("✓ contracts " + contractName)
                 /** constructor - if known */
-                if (insights.contracts[contractName].constructor)
-                    symbols.push(getSymbolForNode(document, insights.contracts[contractName].functions[functionName]._node, vscode.SymbolKind.Constructor))
+                if (insights.contracts[contractName].constructor){
+                    topLevelNode.children.push(astNodeAsDocumentSymbol(document, insights.contracts[contractName].functions[functionName]._node, vscode.SymbolKind.Constructor))
+                }
+                console.log("✓ constructor")
                 /** stateVars */
                 for (var svar in insights.contracts[contractName].stateVars){
-                    symbols.push(getSymbolForNode(document, insights.contracts[contractName].stateVars[svar], vscode.SymbolKind.Field))
-                } 
-
-                for (var name in insights.contracts[contractName].enums){
-                    symbols.push(getSymbolForNode(document, insights.contracts[contractName].enums[name], vscode.SymbolKind.Enum))
-                }
-                for (var name in insights.contracts[contractName].structs){
-                    symbols.push(getSymbolForNode(document, insights.contracts[contractName].structs[name], vscode.SymbolKind.Struct))
-                } 
+                    let symbolAnnotation = getSymbolKindForDeclaration(insights.contracts[contractName].stateVars[svar])
+                    
+                    topLevelNode.children.push(astNodeAsDocumentSymbol(
+                        document, 
+                        insights.contracts[contractName].stateVars[svar], 
+                        symbolAnnotation.symbol,
+                        symbolAnnotation.prefix + symbolAnnotation.name + symbolAnnotation.suffix,
+                        symbolAnnotation.details))
                 
-
+                } 
+                console.log("✓ statevars")
+                
+                for (var name in insights.contracts[contractName].enums){
+                    topLevelNode.children.push(astNodeAsDocumentSymbol(
+                        document, 
+                        insights.contracts[contractName].enums[name], 
+                        vscode.SymbolKind.Enum))
+                }
+                console.log("✓ enums")
+                for (var name in insights.contracts[contractName].structs){
+                    topLevelNode.children.push(astNodeAsDocumentSymbol(document, insights.contracts[contractName].structs[name], vscode.SymbolKind.Struct))
+                } 
+                console.log("✓ structs")
+                var functionLevelNode;
                 /** functions - may include constructor / fallback */
                 for (var functionName in insights.contracts[contractName].functions){
-
-                    //console.log("--->" + functionName + " " + insights.contracts[contractName].functions[functionName].complexity)
-
-                    let prefix = "";
-                    if (solidityVAConfig.outline.decorations){
-                        prefix += getVisibilityToIcon(insights.contracts[contractName].functions[functionName]._node.visibility)
-                        prefix += getStateMutabilityToIcon(insights.contracts[contractName].functions[functionName]._node.stateMutability)
-                    }
-                    let suffix = "    ( ";
-                    suffix += " complex: "+insights.contracts[contractName].functions[functionName].complexity
-                    suffix += " state: " + (insights.contracts[contractName].functions[functionName].accesses_svar?"☑":"☐")
-                    suffix += " )"
-
-                    if(insights.contracts[contractName].functions[functionName]._node.name==null || functionName=="_constructor_" || insights.contracts[contractName].functions[functionName]._node.isConstructor){
-                        symbols.push(getSymbolForNode(document, insights.contracts[contractName].functions[functionName]._node, vscode.SymbolKind.Constructor, "⚜ "+ prefix + "constructor" + suffix))
-                    } else
-                        symbols.push(getSymbolForNode(document, insights.contracts[contractName].functions[functionName]._node, vscode.SymbolKind.Function, prefix + insights.contracts[contractName].functions[functionName]._node.name + suffix))
+                    
+                    let symbolAnnotation = getSymbolKindForDeclaration(insights.contracts[contractName].functions[functionName])
+                    functionLevelNode = astNodeAsDocumentSymbol(
+                        document, 
+                        insights.contracts[contractName].functions[functionName]._node, 
+                        symbolAnnotation.symbol, 
+                        symbolAnnotation.prefix + symbolAnnotation.name + symbolAnnotation.suffix,
+                        symbolAnnotation.details)
+    
+                    topLevelNode.children.push(functionLevelNode)
                     //get all declarations in function
+                    
                     for (var declaration in insights.contracts[contractName].functions[functionName].declarations){
                         let vardec = insights.contracts[contractName].functions[functionName].declarations[declaration];
                         if(declaration=="null")
                             continue
-                        symbols.push(getSymbolForNode(document, vardec, vscode.SymbolKind.Variable))
+                        
+                        let symbolAnnotation = getSymbolKindForDeclaration(vardec)
+                        
+                        functionLevelNode.children.push(astNodeAsDocumentSymbol(
+                            document, 
+                            vardec, 
+                            symbolAnnotation.symbol,
+                            symbolAnnotation.prefix + symbolAnnotation.name + symbolAnnotation.suffix,
+                            symbolAnnotation.details
+                            ))
+                        
                     }
+                    
                 }
+                console.log("✓ functions")
                 
                 /** modifiers */
                 for (var functionName in insights.contracts[contractName].modifiers){
+                    /** 
                     let prefix = "";
                     if (solidityVAConfig.outline.decorations){
                         prefix += getVisibilityToIcon(insights.contracts[contractName].modifiers[functionName].visibility)
                         prefix += getStateMutabilityToIcon(insights.contracts[contractName].modifiers[functionName].stateMutability)
                     }
 
-                    symbols.push(getSymbolForNode(document, insights.contracts[contractName].modifiers[functionName]._node, vscode.SymbolKind.Method, "Ⓜ "+ prefix + insights.contracts[contractName].modifiers[functionName]._node.name))
+                    functionLevelNode = astNodeAsDocumentSymbol(document, insights.contracts[contractName].modifiers[functionName]._node, vscode.SymbolKind.Method, "Ⓜ "+ prefix + insights.contracts[contractName].modifiers[functionName]._node.name)
+                    topLevelNode.children.push(functionLevelNode)
+
+                    */
+
+                    let symbolAnnotation = getSymbolKindForDeclaration(insights.contracts[contractName].modifiers[functionName])
+                    functionLevelNode = astNodeAsDocumentSymbol(
+                        document, 
+                        insights.contracts[contractName].modifiers[functionName]._node, 
+                        symbolAnnotation.symbol, 
+                        symbolAnnotation.prefix + symbolAnnotation.name + symbolAnnotation.suffix,
+                        symbolAnnotation.details)
+
+                    topLevelNode.children.push(functionLevelNode)
                     //get all declarations in function
                     for (var declaration in insights.contracts[contractName].modifiers[functionName].declarations){
                         let vardec = insights.contracts[contractName].modifiers[functionName].declarations[declaration];
                         if(declaration=="null")
                             continue
-                        symbols.push(getSymbolForNode(document, vardec, vscode.SymbolKind.Variable))
+                        let symbolAnnotation = getSymbolKindForDeclaration(vardec)
+                        functionLevelNode.children.push(astNodeAsDocumentSymbol(
+                            document, 
+                            vardec, 
+                            symbolAnnotation.symbol,
+                            symbolAnnotation.prefix + symbolAnnotation.name + symbolAnnotation.suffix,
+                            symbolAnnotation.details
+                            ))
+                
                     }
                 }
+                console.log("✓ modifiers")
                 /** events */
                 for (var functionName in insights.contracts[contractName].events){
+                    /*
                     let prefix = "";
                     if (solidityVAConfig.outline.decorations){
                         prefix += getVisibilityToIcon(insights.contracts[contractName].events[functionName].visibility)
                         prefix += getStateMutabilityToIcon(insights.contracts[contractName].events[functionName].stateMutability)
                     }
-                    symbols.push(getSymbolForNode(document, insights.contracts[contractName].events[functionName], vscode.SymbolKind.Method, "📢 "+ prefix +insights.contracts[contractName].events[functionName].name))
-                }
+                    */
+                    let symbolAnnotation = getSymbolKindForDeclaration(insights.contracts[contractName].events[functionName])
+                    functionLevelNode = astNodeAsDocumentSymbol(
+                        document, 
+                        insights.contracts[contractName].events[functionName]._node, 
+                        symbolAnnotation.symbol, 
+                        symbolAnnotation.prefix + symbolAnnotation.name + symbolAnnotation.suffix,
+                        symbolAnnotation.details)
 
-                 /** functions - may include constructor / fallback */
-                if(solidityVAConfig.outline.inheritance.show){
-                    for (var name in insights.contracts[contractName].inherited_names){
-                        //skip self
-                        if(insights.contracts[contractName].inherited_names[name]==contractName)
+                    topLevelNode.children.push(functionLevelNode)
+
+                    //get all declarations in function
+                    for (var declaration in insights.contracts[contractName].events[functionName].declarations){
+                        let vardec = insights.contracts[contractName].events[functionName].declarations[declaration];
+                        if(declaration=="null")
                             continue
-                        let varname=insights.contracts[contractName].inherited_names[name]+"."+name;
-                        symbols.push(getSymbolForNode(document, getFakeNode(varname, 1), vscode.SymbolKind.Variable, "  ↖"+ varname))
+                        let symbolAnnotation = getSymbolKindForDeclaration(vardec)
+                        functionLevelNode.children.push(astNodeAsDocumentSymbol(
+                            document, 
+                            vardec, 
+                            symbolAnnotation.symbol,
+                            symbolAnnotation.prefix + symbolAnnotation.name + symbolAnnotation.suffix,
+                            symbolAnnotation.details
+                            ))
                     }
                 }
+                console.log("✓ events")
+                 /** functions - may include constructor / fallback */
+                if(solidityVAConfig.outline.inheritance.show){
+
+                    var inheritedLevelNode = astNodeAsDocumentSymbol(
+                        document, 
+                        getFakeNode("↖ ...", 1),
+                        vscode.SymbolKind.Namespace,
+                        )
+                    topLevelNode.children.push(inheritedLevelNode)
+
+                    let contractNodes = {}
+                    for (var name in insights.contracts[contractName].inherited_names){
+                        //skip self
+                        let inheritedFromContractName = insights.contracts[contractName].inherited_names[name];
+                        if(inheritedFromContractName==contractName)
+                            continue
+
+                        let currentContractNode = contractNodes[inheritedFromContractName]
+                        if(typeof currentContractNode=="undefined"){
+                            currentContractNode = astNodeAsDocumentSymbol(
+                                document, 
+                                getFakeNode(inheritedFromContractName, 1), 
+                                vscode.SymbolKind.Class,
+                                "  ↖ "+ inheritedFromContractName)
+                            contractNodes[inheritedFromContractName] = currentContractNode
+                            inheritedLevelNode.children.push(currentContractNode)
+                            
+                        }
+                        // get the item to calculate range/location
+                        let varSymbol = getSymbolKindForDeclaration(g_contracts[inheritedFromContractName].names[name])
+
+                        let inheritanceNode = astNodeAsDocumentSymbol(
+                            document, 
+                            getFakeNode(varSymbol.name, 1), 
+                            varSymbol.symbol, 
+                            "  ↖ "+ varSymbol.prefix + varSymbol.name + varSymbol.suffix,
+                            varSymbol.details)
+                        currentContractNode.children.push(inheritanceNode)
+                    }
+                    
+                }
+                console.log("✓ inheritance")
             }
+            if(token.isCancellationRequested){
+                reject(token)
+                return
+            }
+            console.log("done preparing symbols for: "+ document.fileName)
+            console.log("resolve")
             resolve(symbols);
         });
 
@@ -663,10 +1111,7 @@ function onActivate(context) {
         context.subscriptions.push(
             vscode.languages.reg
         )
-        /** experimental */
-        context.subscriptions.push(
-            vscode.languages.registerDocumentSymbolProvider({language: type}, new SolidityDocumentSymbolProvider()
-        ));
+        
         /** module init */
         onInitModules(context, type);
 
@@ -684,8 +1129,14 @@ function onActivate(context) {
             }
         }, null, context.subscriptions);
 
+        /** experimental */
+        context.subscriptions.push(
+            vscode.languages.registerDocumentSymbolProvider({language: type}, new SolidityDocumentSymbolProvider()
+        ));
+
     }
     
+    setTimeout(onDidChange, 10000);
 }
 
 /* exports */