+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to the "solidity-visual-auditor" extension will be documented in this file.
 
+## v0.0.10
+
+- proper parsing of imports
+- linearization of inheritance
+- highlighting of inherited statevars/methods with location
+- outline view now shows pragmas/imports and inheritance
+- more annotations for outline view
+
 ## v0.0.9
 
 - added solidity parser